Refactor: Outsource contributor organisation event handler to its own module
- Moved contributor organisation logic into a dedicated file (contributor-organisation.js)
- Added full JSDoc documentation for clarity and maintainability
- Refactored role and affiliation input fields for improved structure and readability.
